requirements.txt
numpy
matplotlib
tk
pydicom
pyyaml
pydicom
pillow
opencv-python
torch
cupy-cuda12x
